+++ dashboards/customized/influxdb_v3/plugin/config.py
@@ -3,9 +3,9 @@ from datetime import datetime, timezone
 from enum import Enum
 from typing import Dict, List, assert_never
 
-from .dates import parse_iso_date
-from .window import Window
-from .window_roller import HourlyRoller, MonthlyRoller, roll_windows
+from .utils.dates import parse_iso_date
+from .utils.window import Window
+from .utils.window_roller import HourlyRoller, MonthlyRoller, roll_windows
 
 
 class Mode(Enum):
