+++ dashboards/customized/influxdb_v3/plugin/utils/window.py
